Altium Project for a custom robot sheild
This shield is based on an ATMEGA328P-AU microcontroller that has on board programmer and could be programmed by Arduino IDE (After burning Arduino bootloader through the ISP interface). A 12 volt batter consider as the power supply for the shield and the remaining charge of the battery can be measured by the MCU through software code. This shield also has 5 push buttons , an interface for connecting an 0.96in OLED and a buzzer.